Rename custom venv export command and add usability tweaks
- Uses a positional argument instead of a named arg
- More helpful output
